requirements.txt
discord.py==2.3.2
py-cord==2.4.1
python-dotenv==1.0.0
bcrypt==4.1.2
cherrypy==18.9.0